requirements.txt
absl-py @ file:///C:/ci/absl-py_1615411229697/work
aiohttp @ file:///C:/ci/aiohttp_1614361024229/work
altair==4.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall==0.2.0
base58==2.1.0
black==21.7b0
bleach==3.3.1
blinker==1.4
blis==0.7.4
breadability==0.1.20
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
catalogue==2.0.4
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1613247279197/work
chardet @ file:///C:/ci/chardet_1605303225733/work
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
colorama==0.4.4
coverage @ file:///C:/ci/coverage_1614615074147/work
cryptography @ file:///C:/ci/cryptography_1616769344312/work
cycler==0.10.0
cymem==2.0.5
Cython==0.29.21
debugpy==1.4.0
decorator==5.0.9
defusedxml==0.7.1
docopt==0.6.2
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
entrypoints==0.3
future==0.18.2
gast @ file:///tmp/build/80754af9/gast_1597433534803/work
gensim==4.0.1
gensim-sum-ext==0.1.2
gitdb==4.0.7
GitPython==3.1.18
google-auth @ file:///tmp/build/80754af9/google-auth_1619477128713/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
google-pasta==0.2.0
grpcio @ file:///C:/ci/grpcio_1614884412260/work
h5py==2.10.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1617877484576/work
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver==1.3.1
lxml==4.6.3
mando==0.6.4
Markdown @ file:///C:/ci/markdown_1614364121613/work
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.0
mkl-random @ file:///C:/ci/mkl_random_1618854156666/work
mkl-service==2.3.0
multidict @ file:///C:/ci/multidict_1607362065515/work
murmurhash==1.0.5
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
neattext==0.1.0
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.4.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.1.0
packaging==21.0
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.9.0
pathy==0.6.0
pdfminer.six==20200517
pdfplumber==0.5.28
pickleshare==0.7.5
Pillow==8.3.1
plotly==5.1.0
preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.14.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycorenlp==0.3.0
pycountry==20.7.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pycryptodome==3.10.1
pydantic==1.8.2
pydeck==0.6.2
Pygments==2.9.0
PyJWT==1.7.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
pyPdf==1.13
PyPDF2==1.26.0
pyreadline==2.1
pyrsistent==0.18.0
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
pywinpty==1.1.3
pyzmq==22.1.0
radon==5.0.1
regex==2021.7.6
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
rouge==1.0.1
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-learn==0.23.2
scipy @ file:///C:/ci/scipy_1618856128765/work
seaborn==0.11.1
Send2Trash==1.7.1
six @ file:///C:/ci/six_1605187374963/work
sklearn==0.0
smart-open==5.1.0
smmap==4.0.0
sortedcontainers==2.4.0
spacy==3.1.1
spacy-legacy==3.0.8
srsly==2.4.1
streamlit==0.84.2
sumy==0.8.1
tenacity==8.0.1
tensorboard @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_te/tensorboard_1614593728657/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorflow==2.3.0
tensorflow-estimator @ file:///tmp/build/80754af9/tensorflow-estimator_1599136169057/work/whl_temp/tensorflow_estimator-2.3.0-py2.py3-none-any.whl
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
textblob==0.15.3
thinc==8.0.8
threadpoolctl==2.2.0
toml==0.10.2
tomli==1.1.0
toolz==0.11.1
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
typer==0.3.2
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1611751222202/work
tzlocal==2.1
urllib3 @ file:///tmp/build/80754af9/urllib3_1615837158687/work
validators==0.18.2
vulture==2.3
Wand==0.6.6
wasabi==0.8.2
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
widgetsnbextension==3.5.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
wordcloud==1.8.1
wrapt==1.12.1
yarl @ file:///C:/ci/yarl_1606940076464/work
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
